ARM: configs: multi_v5: Enable Allwinner F1C100
The Kconfig symbols required for the Allwinner F1C100 (MACH_SUNIV) are
currently not selected by any defconfig. sunxi_defconfig only covers the
v7 SoCs, but the F1C100s is ARMv5, so we cannot share a single image.

Add the required symbols to multi_v5_defconfig, to give people some sane
default config when playing around with this chip. This is probably more
important as there are surely not many distros out there supporting
ARMv5 out of the box.

This allows my LicheePi Nano board to boot to a busybox prompt.

The zImage size grows by about 50 KB, in detail:

    text    data     bss     dec     hex   filename
10269442  4284276  678316 15232034 e86c22 vmlinux-old
10348026  4351912  681516 15381454 eab3ce vmlinux-new

14560300			arch/arm/boot/Image-old
14712940			arch/arm/boot/Image-new
6231192				arch/arm/boot/zImage-old
6282928				arch/arm/boot/zImage-new

Signed-off-by: Andre Przywara <andre.przywara@arm.com>