nvme: add NVME_CTRL_MARGINAL flag
Add a new controller flag, NVME_CTRL_MARGINAL, to help multipath I/O
policies to react to a path that is set to a "marginal" state.

The flag is cleared on controller reset, which is often the case when
faulty cabling or transceiver hardware is replaced.

Signed-off-by: Bryan Gurney <bgurney@redhat.com>
---
V10 -> V1:
 - No change
V1 -> V2:
 - Add nvme_ctrl_assign_marginal
 - Fix nvme_fc_ctrl_connectivity_loss clearing wrong bit.